Fix -Wconversion warnings in shader tests.
Bug: 21161908

+++ modules/gles2/functional/es2fShaderLoopTests.cpp
@@ -390,7 +390,7 @@ static ShaderLoopCase* createGenericLoopCase (Context& context, const char* case
 	else
 	{
 		if (loopCountType == LOOPCOUNT_CONSTANT)
-			incrementStr = string("ndx += ") + de::toString(1.0f / numLoopIters);
+			incrementStr = string("ndx += ") + de::toString(1.0f / (float)numLoopIters);
 		else if (loopCountType == LOOPCOUNT_UNIFORM)
 			incrementStr = string("ndx += ") + getFloatFractionUniformName(numLoopIters);
 		else if (loopCountType == LOOPCOUNT_DYNAMIC)
